add BuildConfig.cs and Nunit test case

+++ FastCodeZoo/BuildConfig.cs
@@ -6,6 +6,6 @@ namespace FastCodeZoo
         public const string NameSpace = "FastCodeZoo";
         public const string VersionName = "0.0.1";
         public const int VersionCode = 1;
-        public const int BuildTime = 1599032919;
+        public const long BuildTime = 1599032919;
     }
 }
